Moved classes for day 1
